- move fast filters and indexing to in-memory store backend - move db/common to datastore/common - move saveResolvedNodes to node-model.js (singe usage) - move various node utils from redux/nodes to utils/nodes
