+++ packages/gatsby/src/redux/actions/public.js
@@ -8,7 +8,7 @@ const path = require(`path`)
 const { trueCasePathSync } = require(`true-case-path`)
 const url = require(`url`)
 const { slash, createContentDigest } = require(`gatsby-core-utils`)
-const { hasNodeChanged } = require(`../../redux/nodes`)
+const { hasNodeChanged } = require(`../../utils/nodes`)
 const { getNode } = require(`../../datastore`)
 const sanitizeNode = require(`../../db/sanitize-node`)
 const { store } = require(`..`)
