+++ packages/gatsby/src/schema/__tests__/run-query.js
@@ -1,7 +1,9 @@
 const { store } = require(`../../redux`)
 const { actions } = require(`../../redux/actions`)
 const { getDataStore } = require(`../../datastore`)
-const { runFastFiltersAndSort } = require(`../../redux/run-fast-filters`)
+const {
+  runFastFiltersAndSort,
+} = require(`../../datastore/in-memory/run-fast-filters`)
 
 const makeNodesUneven = () => [
   // Note: This is assumed to be an uneven node count
