+++ packages/gatsby/src/utils/api-runner-node.js
@@ -24,7 +24,7 @@ import {
 } from "../schema/types/type-builders"
 const { emitter, store } = require(`../redux`)
 const { getNodes, getNode, getNodesByType } = require(`../datastore`)
-const { getNodeAndSavePathDependency } = require(`../redux/nodes`)
+const { getNodeAndSavePathDependency } = require(`./nodes`)
 const { getPublicPath } = require(`./get-public-path`)
 const { getNonGatsbyCodeFrameFormatted } = require(`./stack-trace-utils`)
 const { trackBuildError, decorateEvent } = require(`gatsby-telemetry`)
